// frontend/src/redux/actions/types/BookActionType.ts
export enum BookActionType {
    VIEW_BOOK = "VIEW_BOOK",
    VIEW_BOOK_SUCCES = "VIEW_BOOK_SUCCES",
    VIEW_BOOK_NOT_FOUND = "VIEW_BOOK_NOT_FOUND",
    LIST_BOOKS = "LIST_BOOKS",
    LIST_BOOKS_SUCCES = "LIST_BOOKS_SUCCES",
    LIST_BOOKS_CACHED = "LIST_BOOKS_CACHED",
    SEARCH = "SEARCH",
    SEARCH_SUCCESS = "SEARCH_SUCCESS",
    SEARCH_NOT_FOUND = "SEARCH_NOT_FOUND",
    RESET_SEARCH = "RESET_SEARCH"
}